fix(protocol): migrate stream_epoch/seq from u64 to i64
Eliminates unsafe as-casts at every storage boundary. Postgres BIGINT
and SQLite BIGINT are signed i64; the protocol now matches directly.
Values will never exceed i64::MAX in practice.

+++ services/server/src/http/stream_epoch_races.rs
@@ -220,7 +220,7 @@ pub async fn activate_next_stream_epoch_for_race(
         session_id: String::new(),
         forwarder_id,
         reader_ip,
-        new_stream_epoch: next_epoch as u64,
+        new_stream_epoch: next_epoch,
     };
 
     match tokio::time::timeout(
